+++ Source/Executors/Sequential.cs
@@ -3,11 +3,11 @@ using System;
 using System.Threading.Tasks;
 
 
-namespace DotSQL.Executors {
+namespace DotSQL.Executors.Sequential {
     namespace Interfaces {
         internal interface IExecutor {
-            public Core.Result Execute(String query);
-            public Task<Core.Result> ExecuteAsync(String query);
+            public Core.Results.SequentialResult Execute(String query);
+            public Task<Core.Results.SequentialResult> ExecuteAsync(String query);
 
             public System.Data.Common.DbConnection RawConnection();
         }
@@ -16,11 +16,11 @@ namespace DotSQL.Executors {
     internal class SqliteExecutor: Interfaces.IExecutor {
         private System.Data.SQLite.SQLiteConnection Con;
 
-        public SqliteExecutor(Builder.SqliteBuilder builder) {
+        public SqliteExecutor(Builders.Sequential.SqliteBuilder builder) {
             this.Con = new System.Data.SQLite.SQLiteConnection(builder.ConnectionString);
         }
 
-        public Core.Result Execute(String query) {
+        public Core.Results.SequentialResult Execute(String query) {
             try {
                 this.Con.Open();
             }
@@ -31,14 +31,14 @@ namespace DotSQL.Executors {
             var cmd = this.Con.CreateCommand();
             cmd.CommandText = query;
 
-            var result = Core.Result.BuildSqliteResult(cmd.ExecuteReader());
+            var result = Core.Results.SequentialResult.BuildSqliteResult(cmd.ExecuteReader());
 
             this.Con.Close();
 
             return result;
         }
 
-        public async Task<Core.Result> ExecuteAsync(String query) {
+        public async Task<Core.Results.SequentialResult> ExecuteAsync(String query) {
             try {
                 this.Con.Open();
             }
@@ -49,7 +49,7 @@ namespace DotSQL.Executors {
             var cmd = this.Con.CreateCommand();
             cmd.CommandText = query;
 
-            var result = Core.Result.BuildSqliteResult((await cmd.ExecuteReaderAsync()) as System.Data.SQLite.SQLiteDataReader);
+            var result = Core.Results.SequentialResult.BuildSqliteResult((await cmd.ExecuteReaderAsync()) as System.Data.SQLite.SQLiteDataReader);
 
             this.Con.Close();
 
@@ -64,11 +64,11 @@ namespace DotSQL.Executors {
     internal class MySqlExecutor: Interfaces.IExecutor {
         private MySqlConnector.MySqlConnection Con;
 
-        public MySqlExecutor(Builder.MySqlBuilder builder) {
+        public MySqlExecutor(Builders.Sequential.MySqlBuilder builder) {
             this.Con = new MySqlConnector.MySqlConnection(builder.ConnectionString);
         }
 
-        public Core.Result Execute(String query) {
+        public Core.Results.SequentialResult Execute(String query) {
             try {
                 this.Con.Open();
             }
@@ -79,14 +79,14 @@ namespace DotSQL.Executors {
             var cmd = this.Con.CreateCommand();
             cmd.CommandText = query;
 
-            var result = Core.Result.BuildMysqlResult(cmd.ExecuteReader());
+            var result = Core.Results.SequentialResult.BuildMysqlResult(cmd.ExecuteReader());
 
             this.Con.Close();
 
             return result;
         }
 
-        public async Task<Core.Result> ExecuteAsync(String query) {
+        public async Task<Core.Results.SequentialResult> ExecuteAsync(String query) {
             try {
                 this.Con.Open();
             }
@@ -97,7 +97,7 @@ namespace DotSQL.Executors {
             var cmd = this.Con.CreateCommand();
             cmd.CommandText = query;
 
-            var result = Core.Result.BuildMysqlResult((await cmd.ExecuteReaderAsync()) as MySqlConnector.MySqlDataReader);
+            var result = Core.Results.SequentialResult.BuildMysqlResult((await cmd.ExecuteReaderAsync()) as MySqlConnector.MySqlDataReader);
 
             this.Con.Close();
 
@@ -112,11 +112,11 @@ namespace DotSQL.Executors {
     internal class MariadbExecutor: Interfaces.IExecutor {
         private MySqlConnector.MySqlConnection Con;
 
-        public MariadbExecutor(Builder.MariadbBuilder builder) {
+        public MariadbExecutor(Builders.Sequential.MariadbBuilder builder) {
             this.Con = new MySqlConnector.MySqlConnection(builder.ConnectionString);
         }
 
-        public Core.Result Execute(String query) {
+        public Core.Results.SequentialResult Execute(String query) {
             try {
                 this.Con.Open();
             }
@@ -127,14 +127,14 @@ namespace DotSQL.Executors {
             var cmd = this.Con.CreateCommand();
             cmd.CommandText = query;
 
-            var result = Core.Result.BuildMariadbResult(cmd.ExecuteReader());
+            var result = Core.Results.SequentialResult.BuildMariadbResult(cmd.ExecuteReader());
 
             this.Con.Close();
 
             return result;
         }
 
-        public async Task<Core.Result> ExecuteAsync(String query) {
+        public async Task<Core.Results.SequentialResult> ExecuteAsync(String query) {
             try {
                 this.Con.Open();
             }
@@ -145,7 +145,7 @@ namespace DotSQL.Executors {
             var cmd = this.Con.CreateCommand();
             cmd.CommandText = query;
 
-            var result = Core.Result.BuildMariadbResult((await cmd.ExecuteReaderAsync()) as MySqlConnector.MySqlDataReader);
+            var result = Core.Results.SequentialResult.BuildMariadbResult((await cmd.ExecuteReaderAsync()) as MySqlConnector.MySqlDataReader);
 
             this.Con.Close();